LayoutTests/fast/frames/seamless/seamless-inherited-document-style.html

<!DOCTYPE html>
<script src="../../js/resources/js-test-pre.js"></script>
<style>
iframe {
    -webkit-rtl-ordering: visual;
    -webkit-user-modify: read-write;
    -webkit-locale: 'en_US';
    writing-mode: lr;
    direction: rtl;
    font: normal normal normal 18px/normal Ahem;
}
</style>
<iframe id="iframe" seamless srcdoc=""></iframe>
<script>
debug("Test that seamless iframes inherit styles from their parent iframe instead of using StyleResolver::styleForDocument defaults.")
window.onload = function () {
    window.iframe = document.getElementById("iframe");
    window.rootElement = iframe.contentDocument.documentElement;
    shouldBeEqualToString("window.getComputedStyle(rootElement)['-webkit-rtl-ordering']", "visual");
    shouldBeEqualToString("window.getComputedStyle(rootElement)['-webkit-user-modify']", "read-write");
    shouldBeEqualToString("window.getComputedStyle(rootElement)['-webkit-locale']", "en_US");
    shouldBeEqualToString("window.getComputedStyle(rootElement)['writing-mode']", "lr");
    shouldBeEqualToString("window.getComputedStyle(rootElement)['direction']", "rtl");
    shouldBeEqualToString("window.getComputedStyle(rootElement)['font']", "normal normal normal 18px/normal Ahem");
}
</script>
